[Refactor] ProcessingTypeDataProvider and connected classes in separate package
